oM classes split into separate files per name
