+++ RAM_oM/RAM_oM.csproj
@@ -61,7 +61,9 @@
     <Compile Include="RAMLoadCases.cs" />
     <Compile Include="Loads\UniformLoadSet.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Results\RAMGravityResults.cs" />
+    <Compile Include="Results\RAMLineGravityLoad.cs" />
+    <Compile Include="Results\RAMPointGravityLoad.cs" />
+    <Compile Include="Results\RAMFactoredEndReactions.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
