+++ RAM_oM/Results/RAMLineGravityLoad.cs
@@ -26,18 +26,27 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using BH.oM.Base;
+using BH.oM.Structure.Results;
 using BH.oM.Geometry;
 
 namespace BH.oM.Adapters.RAM
-{ 
-    public class ContourLoadSet : BHoMObject
+{
+    public class RAMLineGravityLoad : BHoMObject
     {
         /***************************************************/
         /**** Properties                                ****/
         /***************************************************/
 
-        public Polyline Contour { get; set; } = null;
-        public UniformLoadSet UniformLoadSet { get; set; } = null;
+        public int ObjectId { get; set; } = 0;
+        public double dist1 { get; set; } = 0;
+        public double dist2 { get; set; } = 0;
+        public double DL1 { get; set; } = 0;
+        public double DL2 { get; set; } = 0;
+        public double LL1 { get; set; } = 0;
+        public double LL2 { get; set; } = 0;
+        public double PL1 { get; set; } = 0;
+        public double PL2 { get; set; } = 0;
+        public string type { get; set; } = "";
 
         /***************************************************/
     }
